% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{browse_url}
\alias{browse_url}
\title{open multiple urls sequentially}
\usage{
browse_url(urls, t = 0.2)
}
\arguments{
\item{urls}{character.}

\item{t}{numeric scalar. intervel between opening adjacent webpages, unit:
second. The longer, the more likely your webpage would be opened in the
same order as that of \code{urls}}
}
\value{
\code{NULL}
}
\description{
open multiple urls sequentially
}
\details{
the sequence maybe unstable when file number become large

It depends on external command (\code{system('sleep 0.2')}).
}
\section{demo}{
 \code{libzhuoer::browse_url(c('https://www.bing.com/', 'https://zhuoer.netlify.com/'), 0.4)}
}

