% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_gzip}
\alias{check_gzip}
\title{check integrity of gzip files in a directory}
\usage{
check_gzip(path)
}
\arguments{
\item{path}{character. path to the folder(s) containing gzip files}
}
\value{
character. path of broken gzip files
}
\description{
check integrity of gzip files in a directory
}
\examples{
NULL
}
