% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cols_char}
\alias{cols_char}
\title{Set default column type to character for \strong{readr}}
\usage{
cols_char(..., .default = readr::col_character())
}
\arguments{
\item{..., }{.default  passed on to \code{\link[readr:cols]{readr::cols()}}}
}
\description{
Sometimes most column types are character, but \strong{readr} use
\code{\link[readr:parse_guess]{readr::col_guess()}} by default, thus I write this toy function.
}
\examples{
example_csv <- readr::readr_example('mtcars.csv')

# default column type is guessed
readr::read_csv(example_csv, T)

# sometimes we want to set default column type to character.
readr::read_csv(example_csv, T, cols_char())

# in addition, we can explicitly sepcify some colunmns' type by name
readr::read_csv(example_csv, T, cols_char(mpg = 'd'))

# rather, you can also set default column type to other type (NO abbreviation)
readr::read_csv(example_csv, T, cols_char(.default = readr::col_double()))

}
