% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nested_element}
\alias{nested_element}
\title{get nested elements}
\usage{
nested_element(element, get_element = function(e) {     e[sapply(e,
  is.list)] })
}
\arguments{
\item{element}{usually list. the outermost element.}

\item{get_element}{function. recieve an element and return a list of element.}
}
\value{
list of elements.
}
\description{
get an element (usually a list) itself and elments of it, elements of elments of it, ...
}
\details{
though \code{get_element} restrict the type of elemnt (e.g. the default value force element must be a list), the outermost elment isn't affected. The elements of \code{nested_element(list(1))} are all granted to be lists, but the first (and only) elment \code{nested_element(1)} is numeric, NOT list.
}
\section{legacy}{
 when I support \code{max_level}, \code{nested_list(l, fun)} should return all levels as before
}

\examples{
nested_element(list(1, list(1)))

}
