% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print_or_T}
\alias{print_or_T}
\title{print a variable if not empty}
\usage{
print_or_T(x)
}
\arguments{
\item{x}{}
}
\description{
print a variable if not empty, otherwise return \code{T}
}
\details{
extremely useful with \code{testthat::expect_true()}
}
