% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_char_delim}
\alias{read_char_delim}
\alias{read_char_csv}
\alias{read_char_tsv}
\title{read a delimited file, assuming that every column is character}
\usage{
read_char_csv(file)

read_char_tsv(file)
}
\arguments{
\item{file.}{see \code{\link[readr:read_file]{readr::read_file()}}. for literal input, character will be
collapse into a single string, since using large character for \code{file}
parameter may casue 'C stack usage 7969182 is too close to the limit' for
readr::read_*()}

\item{delim.}{string. passed on to \code{fun}}

\item{fun.}{function. usually read_* from readr}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
read a delimited file, assuming that every column is character
}
\details{
the first line must specify column names. It is used to determine
columns number as it is unlikely to contain \code{delim} (cell value may contain
it in quote).
}
\section{to do}{
 unittest
}

\examples{
read_char_csv("a,b\\n1.0,2.0")
read_char_tsv("a\\tb\\n1.0\\t2.0")
}
