% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-char.R
\name{rm_problematic_row}
\alias{rm_problematic_row}
\title{remove problematic row after you read_*() a file}
\usage{
rm_problematic_row(data)
}
\arguments{
\item{data}{tibble}
}
\value{
tibble
}
\examples{

}
