% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_gzip}
\alias{check_gzip}
\title{check integrity of \code{.gz} files in a directory}
\usage{
check_gzip(path)
}
\arguments{
\item{path}{character. path to the folder(s) containing \code{.gz} files.}
}
\value{
character. path of broken \code{.gz} files.
}
\description{
check integrity of \code{.gz} files in a directory
}
\examples{
dir(system.file('extdata', package = 'libzhuoer'), full = T)

check_gzip(system.file('extdata', package = 'libzhuoer'))

}
